refactor confidants
