+++ src/confidants/index.ts
@@ -1,6 +1,3 @@
-export * from "./confidantList";
-export * from "./confidantController";
+export * from "./ConfidantImages";
 export * from "./ConfidantInfo";
 export * from "./Confidant";
-import { ConfidantList } from "./confidantList";
-export default ConfidantList;
