--- src/pages/confidant/[confidant].ts
@@ -1,2 +1,0 @@
-import { ConfidantDetails } from "../../confidants";
-export default ConfidantDetails;
