--- src/pages/confidant/index.ts
@@ -1,2 +1,0 @@
-import ConfidantList from "../../confidants";
-export default ConfidantList;
